Выделил формы на отдельную страницу

+++ Третий курс/Первый семестр/Базы данных/Database/php/delete_seats.php
@@ -1,11 +1,29 @@
-<?php
-$db = new PDO('mysql:host=db;dbname=kinos', 'devuser', 'devpass');
-print_r($_POST);
-$delete = $db->prepare("DELETE FROM seats WHERE id = :id");
+<!DOCTYPE html>
+<html lang="RU-ru">
+<head>
+    <title>Места</title>
+</head>
+<body>
 
-if ($delete->execute($_POST)) {
-    echo "Действие было совершено удачно";
+
+<?php
+
+if (!empty($_POST)) {
+    $db = new PDO('mysql:host=db;dbname=kinos', 'devuser', 'devpass');
+    print_r($_POST);
+    $delete = $db->prepare("DELETE FROM seats WHERE id = :id");
+
+    if ($delete->execute($_POST)) {
+        echo "Действие было совершено удачно";
+    } else {
+        echo "Ошибка в запросе";
+    }
 }
-else {
-    echo "Ошибка в запросе";
-}
+?>
+
+<form action="delete_seats.php" method="post">
+    <p>id записи <input type="number" name="id" value="">
+    <p><input type="submit"></p>
+</form>
+</body>
+</html>
